+++ src/std_fs/filelib.rs
@@ -200,10 +200,10 @@ fn fs_file_try_readbytes(luau: &Lua, mut multivalue: LuaMultiValue) -> LuaMultiR
 
     // just copied read_file_into_buffer because i don't want to think of an abstraction for it rn
     let (buffy, result): (Option<mlua::Buffer>, &str) = {
-        let try_truncate_f64 = | f: f64, context: &str | -> LuaResult<i32> {
+        let try_truncate_f64 = | f: f64, context: &str | -> LuaResult<i64> {
             let truncated_f = f.trunc();
             if truncated_f != f {
-                Ok(truncated_f as i32)
+                Ok(truncated_f as i64)
             } else {
                 wrap_err!("{} expected {} to be an integer number, got floating point number", function_name_and_args, context)
             }
@@ -245,7 +245,7 @@ fn fs_file_try_readbytes(luau: &Lua, mut multivalue: LuaMultiValue) -> LuaMultiR
         };
     
         // sanity checks
-        let assert_sign = |n: i32, context: &str | -> LuaResult<u64> {
+        let assert_sign = | n: i64, context: &str | -> LuaResult<u64> {
             if n < 0 {
                 wrap_err!("{}: {} cannot be negative", function_name_and_args, context)
             } else {
